Extract playback-status and HLS-encoding hooks
